+++ tests/components/supernotify/test_archive.py
@@ -61,7 +61,12 @@ async def test_integration_archive(mock_hass: HomeAssistant, debug: bool) -> Non
             reobj = json.loads(blob)
         assert reobj["priority"] == "critical"
         assert reobj["target"] == {"person_id": ["person.bob"]}
-        assert reobj["delivered_envelopes"] == uut.last_notification.delivered_envelopes
+        for outcome in ("delivered_envelopes", "delivered_envelopes", "no_envelopes"):
+            for delivery_name in uut.last_notification.deliveries:
+                assert len(reobj["deliveries"][delivery_name].get(outcome, [])) == len(
+                    uut.last_notification.deliveries[delivery_name].get(outcome, [])
+                )
+
         if debug:
             assert reobj["enabled_scenarios"]["alarming"]["enabled"]
         else: